feat: implement query with sort key and refactor rust dynamo client
