+++ src-tauri/src/common/mod.rs
@@ -1,1 +1,2 @@
 pub mod http_client;
+pub mod json_utils;
