+++ src-tauri/src/dynamo/mod.rs
@@ -1,0 +1,5 @@
+pub mod create_item;
+pub mod describe_table;
+pub mod query_table;
+pub mod scan_table;
+pub mod types;
